requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database (sqlite3 is built into Python)

# Voice Processing
SpeechRecognition==3.10.0
pyttsx3==2.90
openai-whisper==20231117

# Natural Language Processing
nltk==3.8.1
spacy==3.7.2

# Audio Processing
pyaudio==0.2.11
soundfile==0.12.1
librosa==0.10.1

# Web Interface
jinja2==3.1.2
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dateutil==2.8.2
regex==2023.10.3 